requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
colorama==0.4.4
coverage==5.3
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
python-slugify==4.0.1
six==1.15.0
text-unidecode==1.3
toml==0.10.1
